incluindo mapping de Contact

+++ Invoice.Domain/Entities/Contact.cs
@@ -10,7 +10,7 @@ namespace InvoiceApi.Domain.Entities
         public string ContactInfo { get; private set; }        
         public ContactType ContactType { get; private set; }        
         public virtual Client Client { get; private set; }        
-        public int ClientId { get; private set; }
+        public long ClientId { get; private set; }
 
         public Contact()
         {
